frontend/requirements.txt
streamlit==1.51.0
request2==0.2
pillow==12.0.0
plotly==6.5.0
pandas==2.3.3